refactor(summary): rename data-fetching drawers to *DrawerHost
CastDrawer, CommentsDrawer, HistoryDrawer, NotesDrawer, SeasonsDrawer,
TriviaDrawer, VideoDrawer, and AddReviewDrawer all own their data
fetching via use* hooks before rendering <Drawer>. Rename to match the
host-pattern convention introduced with MatchDrawerHost so the host vs
content split is consistent across sections.
